Replace `#=>` with `# =>` [ci skip]
@rafaelfranca, I realized I made a couple of commits to docrails with
the same syntax `#=>` that you corrected in f7c7bcd9 with `# =>`. :sob:

If you'd like, I can push this new commits to docrails to fix *all* the
occurrences of `#=>` in the code documentation once and for all.

I know this is cosmetic change.. but it's also just in the documentation.
Should I ship it or not?